feat: improve file password protection toggle
refactor: move PopoverFileItemAction to separate file
